+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.6.12] 2022-10-08
+## [0.6.12] 2022-10-14
 
 ### Added
 
@@ -30,6 +30,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The serialization of prefabs collection resources was fixed.
 
 - The implementation of auto-tests' registration system was fixed.
+
+- The prefabs registry's implementation was fixed to support spawning entities with non-built in component types. 
 
 ## [0.6.11] 2022-09-24